requirements.txt
scipy>=1.1.0
tensorboard>=2.1.1
tensorflow>=2.1.0
tensorflow-estimator>=2.1.0
tensorflow-gpu>=2.1.0
tensorflow-gpu-estimator>=2.1.0
SoundFile>=0.10.2
sacred==0.7.4
librosa>=0.6.2
numpy>=1.15.0
tqdm>=4.28.1
git+https://github.com/craffel/mir_eval.git
mido>=1.2.9
Pillow>=6.2.0
tensorlayer>=2.2.1